CC MP2 energy working

+++ aux.cc
@@ -3,8 +3,7 @@
 #include "aux.h"
 
 using namespace std;
-
-void gprint(vector<int> &v) {
+void gprint(vector<size_t> &v) {
     cout << "[ ";
     for(int i = 0; i < v.size(); i++) {
         cout << v[i] << " ";
